Extract PeerConnectionManager
