+++ MembraneRTC/src/main/java/org/membraneframework/rtc/PeerConnectionListener.kt
@@ -1,0 +1,9 @@
+package org.membraneframework.rtc
+
+import org.webrtc.IceCandidate
+import org.webrtc.MediaStreamTrack
+
+interface PeerConnectionListener {
+    fun onAddTrack(trackId: String, track: MediaStreamTrack)
+    fun onLocalIceCandidate(candidate: IceCandidate)
+}
